CO-1539 Le widget tree_image ne fonctionne plus en prod
Removing attributes that were available in odoo 8.0 but not anymore in odoo 10.0. We replace it with an attribute that make the preview image a bit bigger. We however lose the thumbnail feature that could be clicked on. This could be implemented by diffing the web_tree_image module code betwenn branch 8.0 and 10.0.