CO-3238 Fix bug with mappings
The mapping_id field of many rows in gmc_action was empty, which led to the bug.

Need to update modules child_compassion, intervention_compassion, sbc_compassion and sponsorship_compassion. The migration sets noupdate to false to those fields and the related xml is now no_update=0, so the value of the missing rows will be updated with the value of the xml.

NB:
- Loading gmc_action.xml is very slow, normal
- Might need to run the update of the 4 modules twice (not sure when the query will be run depending on which module is updated first, didn't want to create 4 migrations)