CO-3378 FIX Support replies aren't set to resolved when replied to
Was introduced by CO-3240. Each time the write method of crm.claim was called with an user_id set, the state was set back to "waiting for support".

I deleted this code which solves the bug, but I'm not sure why it was here so it might introduce another bug. Would be good if Joel can review this.