CO-3708 Update react app to prevent concurrent request of sponsorship
Edited endpoints used by odoo to prevent double sponsorship requests. Changed an error message on react side to inform user that a child is already sponsored instead of displaying that the user is the sponsor. Allow to reuse the error in case of concurrent form post.
Translation update were performed for german by Roma, italian by Laura and english and french by myself.