CS-667 Odoo created false partners
According to David, when a crm.claim was created when the email address was refering to a "Linked Partner", this alternative partner was used instead (which is usually empty).
I couldn't reproduce this problem, on my testing environement the create methods was never called with "Liked Partner", but this code works in case it ever happens.

This commit also fix a problem that was partially fixed in a previous task which added "Wordpress Imports" as the crm.claim partner instead of the actual partner.
The email address : wordpress@compassion.ch or any other address must still be present in the ignored_reporter table.