T0010 Feat: Change the error logging to be more precise.
When generating a communication, if the template in use is set with wrong values, or have a wrong behavior, the communication state is set to failure, and the error message is always “Error in template”.
This error message is not enough and need to contains more details. This error message is visible to the user, when the user encounter it, he report it to the IT team. The IT team then correct the template.

In the same time we change an assert into a raise to clean the code